% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkThemis.R
\name{checkThemis}
\alias{checkThemis}
\title{check ThemisMeasurements Unit data
obtains analysisId 1807 and compares to reference data. Output is in export folder and as data.frame}
\usage{
checkThemis(connectionDetails, cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema, outputFolder)
}
\description{
check ThemisMeasurements Unit data
obtains analysisId 1807 and compares to reference data. Output is in export folder and as data.frame
}
