\name{cramerV.test}
\alias{cramerV.test}
\title{Cramer's V}
\usage{
  cramerV.test(source1.data, source2.data, source1.keys,
    source2.keys, source1.variable, source2.variable)
}
\arguments{
  \item{source1.data}{first data source (data.frame)}

  \item{source2.data}{second data source (data.frame)}

  \item{source1.keys}{keys of \code{source1.data}}

  \item{source2.keys}{keys of \code{source2.data}}

  \item{source1.variable}{name of the categorical variable
  (from \code{source1.data})}

  \item{source2.variable}{name of the categorical variable
  (from \code{source2.data})}
}
\value{
  Cramer's V
}
\description{
  Implementation of Cramer's V
}

