% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDQ.R
\name{createMIADLevelOne}
\alias{createMIADLevelOne}
\title{Level one data about dataset MIAD stands for minimum information about a dataset
percentage measures are used that hide exact patient counts}
\usage{
createMIADLevelOne(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema, outputFolder)
}
\description{
Level one data about dataset MIAD stands for minimum information about a dataset
percentage measures are used that hide exact patient counts
}
