% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{customMeasure}
\alias{customMeasure}
\title{do a custom measure}
\usage{
customMeasure(connectionDetails, connectionDetails2)
}
\arguments{
\item{connectionDetails}{connection}

\item{connectionDetails2}{more study parameters}
}
\description{
do a custom measure
}
