% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{customMeasureSql}
\alias{customMeasureSql}
\title{do a custom measure}
\usage{
customMeasureSql(connectionDetails, connectionDetails2, sql)
}
\arguments{
\item{connectionDetails}{connection}

\item{connectionDetails2}{more study parameters}

\item{sql}{sql code to run}
}
\description{
do a custom measure
}
