% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{dashboardLabThresholds}
\alias{dashboardLabThresholds}
\title{Execute all components of the DataQuality study (resultsDatabaseSchema is where Achilles results are)}
\usage{
dashboardLabThresholds(connectionDetails, connectionDetails2)
}
\arguments{
\item{connectionDetails}{connection}

\item{connectionDetails2}{more study parameters}
}
\description{
Execute all components of the DataQuality study (resultsDatabaseSchema is where Achilles results are)
}
