% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataQuality-package.r
\docType{package}
\name{DataQuality}
\alias{DataQuality}
\alias{DataQuality-package}
\title{DataQuality.}
\description{
Package analyzes data quality of a dataset in OMOP CDM format. It uses pre-computations
made by Achilles.
}
