\name{describe}
\alias{describe}
\title{Concise statistical description of a data source}
\usage{
  describe(...)
}
\arguments{
  \item{...}{arguments passed to
  \code{\link[Hmisc]{describe}}}
}
\value{
  See \code{\link[Hmisc]{describe}}
}
\description{
  Concise statistical description of a data source.
}

