% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{doSelectiveExport}
\alias{doSelectiveExport}
\title{selected Achilles Measures}
\usage{
doSelectiveExport(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema = resultsDatabaseSchema,
  cdmVersion = 5,
  workFolder
)
}
\arguments{
\item{connectionDetails}{connection}

\item{cdmDatabaseSchema}{schema}

\item{resultsDatabaseSchema}{result schema}

\item{oracleTempSchema}{oracle specific}

\item{cdmVersion}{version}

\item{workFolder}{where to work}
}
\description{
selected Achilles Measures
}
