% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{.createConnectionDetails2}
\alias{.createConnectionDetails2}
\title{helper object with more connection details}
\usage{
.createConnectionDetails2(
  cdmDatabaseSchema,
  resultsDatabaseSchema = NULL,
  oracleTempSchema = NULL,
  cdmVersion = "5",
  cohortTable = "cohort",
  workFolder = "c:/temp"
)
}
\arguments{
\item{cdmDatabaseSchema}{schema}

\item{resultsDatabaseSchema}{result schema}

\item{oracleTempSchema}{oracle specific}

\item{cdmVersion}{version}

\item{cohortTable}{cohort table name}

\item{workFolder}{where to work}
}
\description{
helper object with more connection details
}
