% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StudySpecific.R
\name{doTree}
\alias{doTree}
\title{experimental funtion with graphic output}
\usage{
doTree(
  connectionDetails,
  cdmDatabaseSchema,
  resultsDatabaseSchema = cdmDatabaseSchema,
  oracleTempSchema = resultsDatabaseSchema,
  cdmVersion = 5,
  workFolder
)
}
\arguments{
\item{connectionDetails}{connection}

\item{cdmDatabaseSchema}{schema}

\item{resultsDatabaseSchema}{result schema}

\item{oracleTempSchema}{oracle specific}

\item{cdmVersion}{version}

\item{workFolder}{where to work}
}
\description{
experimental funtion with graphic output
}
