% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataquality.R
\name{DQR}
\alias{DQR}
\title{Data Quality Report}
\usage{
DQR(df)
}
\arguments{
\item{df}{Input dataframe}

\item{DQR}{Name of the function}
}
\description{
Provides a tool to assess the quality and various parameters of the dataframes.
}
\examples{
DQR(mtcars)

}
