% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe graphics}
\arguments{
\item{lhs, rhs}{A visualisation and a function to apply to it}
}
\description{
This package  uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
