% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataquality.r
\name{populate_DQ_deatil}
\alias{populate_DQ_deatil}
\title{Load the Dataframe}
\usage{
populate_DQ_deatil(base_df = "", file_name = "")
}
\arguments{
\item{base_df}{Name to the input file}

\item{file_name}{Path to the input file}
}
\value{
Void
}
\description{
This function loads a file as a DF. calculates the stats and save it
DB.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
