\name{quality.report}
\alias{quality.report}
\title{Quality report}
\usage{
  quality.report(src.data, src.vars, src.keys, ref.data,
    ref.vars, ref.keys, check.pattern = NULL,
    output = "report.html")
}
\arguments{
  \item{data}{data}

  \item{output}{name of genereted html document}
}
\description{
  Quality report
}

