% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDQ.R
\name{createMIAD}
\alias{createMIAD}
\title{Level one data about dataset MIAD stands for minimum information about a dataset
percentage measures are used that hide exact patient counts}
\usage{
createMIAD(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema, outputFolder, level = 3)
}
\description{
Level one data about dataset MIAD stands for minimum information about a dataset
percentage measures are used that hide exact patient counts
}
