% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetDQ.R
\name{createMIADLevelOne}
\alias{createMIADLevelOne}
\title{Level one data about dataset}
\usage{
createMIADLevelOne(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema, outputFolder)
}
\description{
Level one data about dataset
}

